SendMessage calls block startup
The app hangs after calling SendMessage with a system colour change message.
According to the docs (https://learn.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-sendmessage):
"Sends the specified message to a window or windows. The SendMessage function calls the window procedure for the specified window and does not return until the window procedure has processed the message."

In the hang we see, we have a single threaded application which has not yet even begun processing windows messages and yet is going to send it, and everyone else, a message which it will never be able to read, so it waits forever as a result.
A similar things is explained here: https://stackoverflow.com/a/1956702/5751627
I'm not clear enough on the original intent to know the aim, so for now I'm going to settle for changing SendMessage to PostMessage which won't wait for the result.