OF_SEARCH bit set prevents file creation
It seems that if the original value of OF_SEARCH (0x0400) is passed to a call to open with create, it will fail regardless, so set this to zero so that it has no effect at all. At this point it's still unclear what the correct answer will be to this, or indeed if it was actually ever needed.