Invalid calls to DeleteDC cause program to exit
In fry_dc the logic first calls ReleaseDC and then DeleteDC. If either fail then they post a quit message which causes the app to exit.
What we see in the debugger is that DeleteDC fails and so the application quits.
DeleteDC docs state:  https://learn.microsoft.com/en-us/windows/win32/api/wingdi/nf-wingdi-deletedc
"An application must not delete a DC whose handle was obtained by calling the GetDC function. Instead, it must call the ReleaseDC function to free the DC."
In every case throughout the code that ultimately calls fry_dc I can trace (and it's quite tricky to do this) GetDC is the call used to get the DC, so the call to DeleteDC should never be called in fry_dc.