More WNDPROC and some DLGPROC signature changes
Similar changes to the previous one in regards WNDPROC signatures, but this time these aren't created Windows procedures they are typically to gain access to an edit controls existing windows procedure or to set a new one.
The rest of the changes are sorting out dialog box procedure signatures, typically for CreateDialog calls.
I added a few casts to the right type for GetWindow* functions for certain messages too.